chore: move HashBuiltinRunner to own file
